- Added `ProfileController` with the following endpoints:   - `GET /profiles`: Retrieves all profiles.   - `GET /profiles/{id}`: Retrieves a profile by ID.   - `POST /profiles`: Adds a new profile.   - `PUT /profiles/{id}`: Updates an existing profile.   - `DELETE /profiles/{id}`: Deletes a profile by ID. - Integrated `ProfileService` - Configured `ProfileRepository`

+++ src/main/java/com/habitoplus/habitoplusback/Service/AccountService.java
@@ -15,7 +15,7 @@ public class AccountService {
         return accountRepository.findAll();
     }
 
-    public Account getAccountById(Long id) {
+    public Account getAccountById(int id) {
         return accountRepository.findById(id).orElse(null);
     }
     public Account getAccountByEmail(String email) {
@@ -34,7 +34,7 @@ public class AccountService {
         return true;
     }
 
-    public boolean deleteAccount(Long id) {
+    public boolean deleteAccount(int id) {
         Account existingAccount = accountRepository.findById(id).orElse(null);
         if (existingAccount != null) {
             existingAccount.setStatus(false);
